:arrow_up: :art: :racehorse: Rewrite getIcon a little bit.
Bump resource commit SHA.

Rewrite getIcon in ItemSteamAxe, ItemSteamDrill, and ItemSteamShovel to allow for multiple upgrades installed at once (2). This involves rendering the core first, then the upgrade. When there is no upgrade, it will render the standard default textures. This allows for one less render pass, since we only render the base tool when we absolutely need to. This should slightly improve performance, but that wasnt the main goal of this commit.