:gem: :wrench: Book API redesign (#470)
New high level classes, BookCategory and BookEntry, which implement BookPiece. A BookPiece is a lockable/unlockable/hidable piece in the book. Pages are not BookPieces, but I might change that. BookCategories can contain BookEntries and subcategory BookCategories. BookEntries can contain BookPages. In the current state, the book functions and appears identically as it did before. Remove a lot of redundant fields in BookPageRegistry, and basically rewrite that entire class. This new stuff allows for pieces in the book to be hidden, like the creative tank, or gated behind doing something.