:package: Greatly improve usability of ResearchRecipes. Remove ResearchObject (and according Item and Block variants) interface. Instead, a TriPredicate is passed to the ResearchRecipe constructor which is tested against in #matches. New TriPredicate functional interface. ResearchRecipe constructor now takes an ItemStack instead of a Block/Item.
This allows for more complex and coherent recipe matching code. Matching code is no longer limited to only unlocked checking. You can now output metadata and amount > 1 items. Lastly, this allows for adding ResearchRecipes for items which the consumer does not have control over (since it does not require the object to implement a certain class).